# openapi.yaml
openapi: 3.1.0
info:
  title: ainu-corpora-api
  version: "1"
  description: >
    Corpus search / KWIC concordance / POS-search API for ainu-corpora.
    Every response is an envelope: { api_version, data } on success,
    { api_version, error: { code, message } } on failure. Routes are under /v1;
    breaking changes bump to /v2. CORS is open (GET only).
servers:
  - url: https://corpus.aynu.org
components:
  schemas:
    Envelope:
      type: object
      required: [api_version]
      properties:
        api_version: { type: string, example: "1" }
        data: {}
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
    CorpusRow:
      type: object
      properties:
        id: { type: string }
        text: { type: string }
        translation: { type: string, nullable: true }
        dialect: { type: string, nullable: true }
        author: { type: string, nullable: true }
        collection: { type: string, nullable: true }
        document: { type: string, nullable: true }
        uri: { type: string, nullable: true }
        source_slug: { type: string, nullable: true, description: "db.aynu.org source-record slug (https://db.aynu.org/sources/<slug>); null when the collection has no registered source" }
    KwicLine:
      type: object
      properties:
        sentence_id: { type: string }
        left: { type: string }
        node: { type: string }
        right: { type: string }
        translation: { type: string, nullable: true }
        dialect: { type: string, nullable: true }
        author: { type: string, nullable: true }
        uri: { type: string, nullable: true }
        source_slug: { type: string, nullable: true, description: "db.aynu.org source-record slug; null when unregistered" }
    PosLine:
      allOf:
        - $ref: "#/components/schemas/KwicLine"
        - type: object
          properties:
            upos: { type: string, nullable: true }
            lemma: { type: string, nullable: true }
    KwicToken:
      type: object
      description: One token in an annotated KWIC window.
      properties:
        i: { type: integer, description: index within the sentence }
        s: { type: string, description: surface (as printed) }
        n: { type: string, description: surface_norm }
        p: { type: string, nullable: true, description: "display POS (morpheme DB override, e.g. PERS, else raw UPOS)" }
        u: { type: string, nullable: true, description: "raw tagger UPOS" }
        l: { type: string, nullable: true, description: lemma }
        x: { type: string, nullable: true, description: xpos (fine tag) }
        f: { type: string, nullable: true, description: morphological features }
        g: { type: string, nullable: true, description: "short English gloss from ainu-morpheme-database" }
        mc: { type: string, nullable: true, description: "morpheme DB category (pers, vt, n, sfx, ...)" }
        cl: { type: integer, description: 1 if clitic }
        node: { type: boolean, description: present+true on the node token }
        alt:
          type: array
          description: >
            Alternate homograph readings of this surface that lost the display slot,
            most-likely first (e.g. pa displays NOUN "head" with alt PL, "mouth").
            Omitted when the surface has no competing reading.
          items:
            type: object
            properties:
              p: { type: string, nullable: true, description: display POS of the alternate reading }
              g: { type: string, description: gloss of the alternate reading }
              mc: { type: string, nullable: true, description: morpheme DB category of the alternate reading }
    KwicAnnotatedLine:
      type: object
      properties:
        sentence_id: { type: string }
        left: { type: array, items: { $ref: "#/components/schemas/KwicToken" } }
        node: { type: array, items: { $ref: "#/components/schemas/KwicToken" } }
        right: { type: array, items: { $ref: "#/components/schemas/KwicToken" } }
        left_text: { type: string }
        node_text: { type: string }
        right_text: { type: string }
        translation: { type: string, nullable: true }
        dialect: { type: string, nullable: true }
        author: { type: string, nullable: true }
        uri: { type: string, nullable: true }
        source_slug: { type: string, nullable: true, description: "db.aynu.org source-record slug; null when unregistered" }
    Collocate:
      type: object
      properties:
        token: { type: string }
        upos: { type: string, nullable: true }
        freq: { type: integer, description: co-occurrences in the window }
        total: { type: integer, description: collocate's total corpus frequency }
        log_dice: { type: number, description: logDice association }
        t_score: { type: number }
        mi: { type: number, description: pointwise mutual information }
paths:
  /health:
    get:
      summary: Liveness probe
      responses: { "200": { description: ok } }
  /v1/search:
    get:
      summary: Full-text corpus search (FTS5 trigram substring)
      parameters:
        - { name: q, in: query, required: true, schema: { type: string } }
        - { name: lang, in: query, schema: { type: string, enum: [ain, jpn, any], default: any } }
        - { name: dialect, in: query, schema: { type: string } }
        - { name: author, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
      responses:
        "200": { description: "CorpusRow[] in data" }
  /v1/concordance:
    get:
      summary: KWIC concordance for a node word (left / node / right)
      parameters:
        - { name: q, in: query, required: true, schema: { type: string }, description: node word (normalized) }
        - { name: window, in: query, schema: { type: integer, default: 40 }, description: context chars each side }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: sort, in: query, schema: { type: string, enum: [none, left, right], default: none } }
        - { name: match, in: query, schema: { type: string, enum: [exact, prefix], default: exact } }
        - { name: dialect, in: query, schema: { type: string } }
        - { name: author, in: query, schema: { type: string } }
      responses:
        "200": { description: "KwicLine[] in data" }
  /v1/pos:
    get:
      summary: POS / grammatical search (machine UD tags; Latin-script only)
      description: >
        Provide at least one of upos/lemma/surface/next_upos/next_surface.
        next_* constrain the immediately following token (idx+1) — a self-join.
        Example: ?upos=VERB&next_surface==an  → intransitive verbs taking =an.
      parameters:
        - { name: upos, in: query, schema: { type: string }, description: "UD UPOS e.g. NOUN, VERB, ADP, PART, ADV, DET, AUX, SCONJ, NUM, PRON, INTJ" }
        - { name: lemma, in: query, schema: { type: string } }
        - { name: surface, in: query, schema: { type: string } }
        - { name: next_upos, in: query, schema: { type: string } }
        - { name: next_surface, in: query, schema: { type: string } }
        - { name: window, in: query, schema: { type: integer, default: 40 } }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: dialect, in: query, schema: { type: string } }
        - { name: author, in: query, schema: { type: string } }
      responses:
        "200": { description: "PosLine[] in data" }
        "400": { description: "error: missing_filter" }
  /v1/kwic:
    get:
      summary: Annotated, token-level KWIC (clickable words, POS/gloss, accent-folding, inflection)
      description: >
        Richer than /v1/concordance: each line carries left/node/right TOKEN
        arrays (surface + upos + lemma + xpos + feats + clitic flag) plus plain
        left_text/node_text/right_text. match=fold (default) is accent-insensitive
        (néa == nea, rámat == ramat). match=regex treats q as a word regex over
        the folded key — case-insensitive, unanchored (use ^ $), e.g. q=ech?i
        matches eci and echi; resolved against the vocabulary (capped at the
        1000 most frequent matching word forms). expand=plural also surfaces a
        verb's plural (e.g. arpa → paye). sort supports l1..l3 / r1..r3 (Nth
        neighbour), node, dialect, author.
      parameters:
        - { name: q, in: query, required: true, schema: { type: string } }
        - { name: ctx, in: query, schema: { type: integer, default: 6 }, description: context tokens each side }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 }, description: "pagination offset; response carries meta:{total,offset,limit}" }
        - { name: match, in: query, schema: { type: string, enum: [fold, exact, prefix, regex], default: fold } }
        - { name: sort, in: query, schema: { type: string, enum: [none, left, right, l1, l2, l3, r1, r2, r3, node, dialect, author], default: none } }
        - { name: expand, in: query, schema: { type: string, enum: [none, plural, all], default: none }, description: fold in inflected relatives (morph_forms) }
        - { name: upos, in: query, schema: { type: string }, description: constrain the node's POS }
        - { name: clitic, in: query, schema: { type: string, enum: [any, only, exclude], default: any } }
        - { name: dialect, in: query, schema: { type: string } }
        - { name: author, in: query, schema: { type: string } }
      responses:
        "200": { description: "KwicAnnotatedLine[] in data" }
        "400": { description: "error: missing_query | bad_regex" }
  /v1/collocation:
    get:
      summary: Collocates of a node word, scored by logDice / t-score / MI
      description: >
        Tokens co-occurring within ±window token positions of the node, with
        association measures. span restricts to one side. PUNCT is excluded.
      parameters:
        - { name: q, in: query, required: true, schema: { type: string } }
        - { name: window, in: query, schema: { type: integer, default: 5 }, description: "± token positions (max 10)" }
        - { name: span, in: query, schema: { type: string, enum: [both, left, right], default: both } }
        - { name: measure, in: query, schema: { type: string, enum: [log_dice, t_score, mi], default: log_dice }, description: sort key }
        - { name: minCount, in: query, schema: { type: integer, default: 3 } }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: dialect, in: query, schema: { type: string } }
        - { name: author, in: query, schema: { type: string } }
      responses:
        "200": { description: "{ node, node_freq, corpus_tokens, collocates: Collocate[] } in data" }
  /v1/structural:
    get:
      summary: Structural (CQL-lite) adjacent token-sequence search
      description: >
        Match a sequence of token constraints via idx+1 self-joins. Pattern
        syntax: [upos=VERB] [surface=arpa] [lemma=arpa] [surface=ku.*] (prefix)
        [] (wildcard); bare word → surface, bare ALLCAPS → upos; & joins
        constraints in one position. /ech?i/ (bare or [surface=/ech?i/]) is a
        word regex over the folded key — case-insensitive, unanchored (use ^ $);
        e.g. pattern=/ech?i/ /^a/ finds eci|echi followed by an a-initial word.
        Max 6 positions. e.g. [upos=NOUN] [upos=NOUN].
      parameters:
        - { name: pattern, in: query, required: true, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: dialect, in: query, schema: { type: string } }
        - { name: author, in: query, schema: { type: string } }
      responses:
        "200": { description: "StructuralLine[] in data (match[] + match_text + left_text/right_text)" }
        "400": { description: "error: missing_pattern | bad_regex" }
  /v1/analytics:
    get:
      summary: Distribution of a node word across dialect/author/collection + POS spread
      parameters:
        - { name: q, in: query, required: true, schema: { type: string } }
        - { name: match, in: query, schema: { type: string, enum: [fold, exact, prefix], default: fold } }
        - { name: top, in: query, schema: { type: integer, default: 10 } }
      responses:
        "200": { description: "{ node, total, dialects[], authors[], collections[], upos[] } in data" }
  /v1/inflections:
    get:
      summary: Inflectional relatives of a word (singular↔plural, possessed forms)
      description: >
        From the ainu-morpheme-database morphology layer (morph_forms). Matches a
        word as either lemma or surface and returns linked forms. Powers the UI
        inflection chooser and expand=plural in /v1/kwic.
      parameters:
        - { name: word, in: query, required: true, schema: { type: string } }
      responses:
        "200": { description: "{ query, fold, forms[] } in data" }
  /v1/examples:
    get:
      summary: Curated, runnable search examples (self-describing tour)
      parameters:
        - { name: mode, in: query, schema: { type: string, enum: [kwic, pos, collocation, structural, analytics, inflection, text] } }
      responses:
        "200": { description: "SearchExample[] in data ({ mode, label, desc, params, path })" }
  /v1/stats:
    get:
      summary: Corpus totals + top dialect distribution
      responses: { "200": { description: "{ sentences, top_dialects } in data" } }
  /v1/meta:
    get:
      summary: Generic precomputed meta value
      parameters: [ { name: key, in: query, required: true, schema: { type: string } } ]
      responses: { "200": { description: "string|null in data" } }
  /v1/freq/word:
    get:
      summary: Frequency + rank for one already-normalized token
      parameters: [ { name: token, in: query, required: true, schema: { type: string } } ]
      responses: { "200": { description: "{ token, found, count, is_stopword, rank } in data" } }
  /v1/freq/list:
    get:
      summary: Most-frequent tokens
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 100 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
        - { name: includeStopwords, in: query, schema: { type: boolean, default: false } }
        - { name: minCount, in: query, schema: { type: integer, default: 1 } }
      responses: { "200": { description: "{ token, count, is_stopword }[] in data" } }
  /v1/stopwords:
    get: { summary: Canonical stopword list, responses: { "200": { description: "string[] in data" } } }
  /v1/stopword:
    get:
      summary: Is one already-normalized token a stopword?
      parameters: [ { name: token, in: query, required: true, schema: { type: string } } ]
      responses: { "200": { description: "{ token, is_stopword } in data" } }
  /v1/candidates:
    get:
      summary: Raw vocab-gap candidates (glossary subtraction stays in the MCP)
      parameters: [ { name: minCount, in: query, schema: { type: integer, default: 1 } } ]
      responses: { "200": { description: "candidate[] in data" } }